mv all inline tests to tests/ dir
