+++ tests/helpers/cleanup_metadb_tables.sql
