--- tests/modules/ses_approval/outputs.tf
@@ -1,7 +1,0 @@
-output "bucket_arn" {
-  value = aws_s3_bucket.this.arn
-}
-
-output "id" {
-  value = aws_s3_bucket.this.id
-}
