+++ tests/unit/tf-module/provider_switches.hcl
